Reset projects.

+++ Tomb3/tomb3/3dsystem/scalespr.h
@@ -4,4 +4,4 @@
 short* ins_room_sprite(short* objptr, long num);
 void S_DrawScreenSprite2d(long x, long y, long z, long scaleH, long scaleV, short sprnum, short shade, ushort flags);
 void S_DrawScreenSprite(long x, long y, long z, long scaleH, long scaleV, short sprnum, long shade, ushort flags);
-void S_DrawSprite(ulong flags, long x, long y, long z, short sprnum, ulong shade, short scale);
+void S_DrawSprite(ulong flags, long x, long y, long z, short sprnum, short shade, short scale);
